Skin-care_Backend/requirements.txt
flask
flask-cors
tensorflow
torch
torchvision
pillow
numpy
transformers
rapidfuzz
python-dotenv
sentencepiece
